+++ src/class-wc-gateway-bitpay.php
@@ -1,0 +1,1 @@
+<?php require_once __DIR__ ."/class-wc-gateway-btcpay.php" ?>